+++ tests/integration/ai/test_hooks.py
@@ -18,16 +18,18 @@ from pydantic import BaseModel, Field
 from splunklib.ai import Agent
 from splunklib.ai.conversation_store import InMemoryStore
 from splunklib.ai.hooks import (
+    after_agent,
+    after_model,
+    before_agent,
+    before_model,
+)
+from splunklib.ai.limits import (
     StepLimitMiddleware,
     StepsLimitExceededException,
     TimeoutExceededException,
     TimeoutLimitMiddleware,
     TokenLimitExceededException,
     TokenLimitMiddleware,
-    after_agent,
-    after_model,
-    before_agent,
-    before_model,
 )
 from splunklib.ai.messages import AgentResponse, AIMessage, HumanMessage
 from splunklib.ai.middleware import (
